+++ app/(search)/actions.ts
@@ -1,18 +1,9 @@
 "use server";
 
-import { dedupeResults } from "@/app/(search)/dedupeResults";
-import { getToken } from "@/app/[company]/services/rues";
-import {
-  RECAPTCHA_SEARCH_ACTION,
-  VALID_RUES_CATEGORIES,
-} from "@/app/shared/lib/constants";
-import { isValidNit } from "@/app/shared/lib/isValidNit";
+import { RECAPTCHA_SEARCH_ACTION } from "@/app/shared/lib/constants";
 import { verifyRecaptcha } from "@/app/shared/lib/verifyRecaptcha";
-import { mapRuesResultToCompanySummary } from "@/app/shared/mappers/mapRuesResultToCompany";
-import { RUES } from "@mauriciorobayo/rues-api";
+import { getSearchResultsByCompanyName } from "@/app/shared/services/rues/api";
 import { unstable_cache } from "next/cache";
-
-const rues = new RUES();
 
 export async function search({
   companyName,
@@ -33,28 +24,11 @@ export async function search({
     return null;
   }
 
-  return getSearchResultsByCompanyName(companyName);
+  return getSearchResultsByCompanyNameCached(companyName);
 }
 
-const getSearchResultsByCompanyName = unstable_cache(
-  async (companyName: string) => {
-    const token = await getToken();
-    const response = await rues.advancedSearch({
-      query: { razon: companyName },
-      token,
-    });
-    if (response.status === "error") {
-      return null;
-    }
-    return dedupeResults(
-      (response.data.registros ?? []).filter(
-        (record) =>
-          isValidNit(Number(record.nit)) &&
-          VALID_RUES_CATEGORIES.includes(record.categoria) &&
-          !!record.id_rm,
-      ),
-    ).map(mapRuesResultToCompanySummary);
-  },
+const getSearchResultsByCompanyNameCached = unstable_cache(
+  getSearchResultsByCompanyName,
   undefined,
   {
     revalidate: 2 * 24 * 60 * 60,
