+++ app/shared/services/rues/processAdvancesSearchResults.ts
@@ -1,3 +1,5 @@
+import { VALID_RUES_CATEGORIES } from "@/app/shared/lib/constants";
+import { isValidNit } from "@/app/shared/lib/isValidNit";
 import type { BusinessRecord } from "@mauriciorobayo/rues-api";
 
 export function dedupeResults<
@@ -14,5 +16,15 @@ export function dedupeResults<
   });
   return Array.from(
     new Map(sortedResults.map((result) => [result.nit, result])).values(),
+  );
+}
+
+export function processAdvancedSearchResults(results: BusinessRecord[]) {
+  return dedupeResults(
+    (results ?? []).filter(
+      (record) =>
+        isValidNit(Number(record.nit)) &&
+        VALID_RUES_CATEGORIES.includes(record.categoria),
+    ),
   );
 }
